+++ app/Http/Requests/StoreOfferRequest.php
@@ -3,6 +3,7 @@
 namespace App\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Support\Facades\Auth;
 
 class StoreOfferRequest extends FormRequest
 {
@@ -13,7 +14,7 @@ class StoreOfferRequest extends FormRequest
      */
     public function authorize()
     {
-        return false;
+        return true;
     }
 
     /**
@@ -24,7 +25,29 @@ class StoreOfferRequest extends FormRequest
     public function rules()
     {
         return [
-            //
+            'user_id' => 'required|uuid',
+            'type' => 'required|in:HOUSING,TRANSPORT_ASSISTANCE',
+            'lat' => 'required|numeric',
+            'lng' => 'required|numeric',
+            'canTakePets' => 'required|boolean',
+            'canTakeSingles' => 'required|boolean',
+            'canTakeCouples' => 'required|boolean',
+            'canTakeFamilies' => 'required|boolean',
         ];
+    }
+
+    /**
+     * Prepare the data for validation.
+     * 
+     * Adding random noise to the point to increase privacy.
+     *
+     * @return void
+     */
+    protected function prepareForValidation()
+    {
+        $this->merge([
+            'lat' => $this->lat + (0.00001 * rand(0, 30)),
+            'lng' => $this->lng + (0.00001 * rand(0, 30)),
+        ]);
     }
 }