+++ app/Models/Offer.php
@@ -4,10 +4,12 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Offer extends Model
 {
     use HasFactory;
+    use SoftDeletes;
 
     protected $fillable = [
         'user_id',
@@ -15,9 +17,10 @@ class Offer extends Model
         'lng',
         'type',
         'notes',
-        'duration',
-        'capacity',
         'canTakePets',
+        'canTakeSingles',
+        'canTakeCouples',
+        'canTakeFamilies',
         'assistanceType',
     ];
 }