+++ app/Models/PointOfInterest.php
@@ -4,12 +4,15 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
 
 class PointOfInterest extends Model
 {
     use HasFactory;
+    use SoftDeletes;
 
     protected $fillable = [
+        'user_id',
         'name',
         'type',
         'lat',