+++ database/seeders/LocationSeeder.php
@@ -27,8 +27,8 @@ class LocationSeeder extends Seeder
                 'stateName' => $data[5],
                 'type' => $data[6],
                 'localGovernmentArea' => $data[7],
-                'lng' => $data[13],
-                'lat' => $data[14],
+                'lat' => $data[13],
+                'lng' => $data[14],
                 'timezone' => $data[15],
             ]);
         }