+++ resources/js/components/sidebar-layout.tsx
@@ -1,18 +1,58 @@
-import React, { Fragment, useState } from 'react'
+import React, { Fragment, useState, useCallback, useRef, useEffect } from 'react'
 import { Dialog, Transition } from '@headlessui/react'
 import {
     Bars3Icon,
     UsersIcon,
     XMarkIcon,
-    PlusIcon
+    PlusIcon,
+    MapPinIcon,
 } from '@heroicons/react/24/outline'
-import { usePage } from '@inertiajs/inertia-react'
+import {
+    HomeIcon,
+    ExclamationTriangleIcon,
+    PencilIcon,
+    TrashIcon,
+    PencilSquareIcon,
+    CheckCircleIcon,
+    PlusCircleIcon
+} from '@heroicons/react/24/solid'
+import { usePage, useForm } from '@inertiajs/inertia-react'
 import { Inertia } from '@inertiajs/inertia'
+import Map, { GeolocateControl, Marker } from 'react-map-gl';
 
-const CreateModal = ({ open, toggleFn }) => {
+const CreateOfferModal = ({ open, toggleFn }) => {
+    const mapRef = useRef();
+    const { auth } = usePage().props
+    const { data, setData, post, processing, errors } = useForm({
+        canTakeSingles: false,
+        canTakeCouples: false,
+        canTakeFamilies: false,
+        canTakePets: false,
+        lat: -37.8136,
+        lng: 144.9631,
+        type: 'HOUSING',
+        user_id: auth.user.id
+    })
+
+    const onMapLoad = useCallback(() => {
+        mapRef.current.on('move', () => {
+            const coordinates = mapRef.current.getCenter();
+
+            setData(data => ({ ...data, lat: coordinates.lat }));
+            setData(data => ({ ...data, lng: coordinates.lng }));
+        });
+    }, []);
+
+    function submit(e) {
+        e.preventDefault()
+        post('/api/offers', {
+            onSuccess: () => toggleFn(false)
+        })
+    }
+
     return (
         <Transition.Root show={open} as={Fragment}>
-            <Dialog as="div" className="relative z-10" onClose={toggleFn}>
+            <Dialog as="div" className="relative z-10" onClose={() => toggleFn(false)}>
                 <Transition.Child
                     as={Fragment}
                     enter="ease-out duration-300"
@@ -37,32 +77,294 @@ const CreateModal = ({ open, toggleFn }) => {
                             leaveTo="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
                         >
                             <Dialog.Panel className="relative transform overflow-hidden rounded-lg bg-white px-4 pt-5 pb-4 text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-sm sm:p-6">
-                                <div>
-                                    <div className="mx-auto flex h-12 w-12 items-center justify-center rounded-full bg-green-100">
-                                        <PlusIcon className="h-6 w-6 text-green-600" aria-hidden="true" />
-                                    </div>
-                                    <div className="mt-3 text-center sm:mt-5">
-                                        <Dialog.Title as="h3" className="text-lg font-medium leading-6 text-gray-900">
-                                            Create an Offer or Report
-                                        </Dialog.Title>
-                                        <div className="mt-2">
-                                            <p className="text-sm text-gray-500">
-                                                If you can provide temporary emergency accomodation or report a danger like road damage, please report it here.
+                                <form onSubmit={submit} className="space-y-6 divide-y divide-gray-200">
+                                    <div className="space-y-6 divide-y divide-gray-200">
+                                        <div>
+                                            <h3 className="text-lg font-medium leading-6 text-gray-900">Offer</h3>
+                                            <p className="mt-1 text-sm text-gray-500">
+                                                This information will be displayed publicly so be careful what you share.
                                             </p>
                                         </div>
-                                    </div>
-                                </div>
-                                <div className="mt-5 sm:mt-6">
-                                    {/* TODO: add here */}
 
-                                    <button
-                                        type="button"
-                                        className="inline-flex w-full justify-center rounded-md border border-transparent bg-indigo-600 px-4 py-2 text-base font-medium text-white shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 sm:text-sm"
-                                        onClick={() => toggleFn(false)}
-                                    >
-                                        Go back to dashboard
-                                    </button>
-                                </div>
+                                        <div className="pt-6">
+                                            <legend className="sr-only">Location</legend>
+                                            <div className="text-base font-medium text-gray-900" aria-hidden="true">
+                                                Location
+                                            </div>
+                                            <div id='offer-location-select' className="h-64 w-full relative mt-4">
+                                                <Map
+                                                    ref={mapRef}
+                                                    initialViewState={{
+                                                        longitude: 144.9631,
+                                                        latitude: -37.8136,
+                                                        zoom: 5
+                                                    }}
+                                                    style={{
+                                                        width: '100%',
+                                                        height: '100%'
+                                                    }}
+                                                    onLoad={onMapLoad}
+                                                    mapStyle="mapbox://styles/mapbox/streets-v9"
+                                                    mapboxAccessToken="pk.eyJ1IjoiZGFuaWVsZmVyZ3Vzb24iLCJhIjoiY2w5YXFjazNtMGp1ZTNwcXdtMjBlYTc2YyJ9.2Cz8UmqgWB4VpagnJ6_ATw"
+                                                >
+                                                    <Marker longitude={data.lng} latitude={data.lat} anchor="bottom" />
+                                                    <GeolocateControl />
+                                                </Map>
+                                            </div>
+                                        </div>
+
+                                        <div className="pt-6">
+                                            {/* Capacity */}
+                                            <fieldset>
+                                                <legend className="sr-only">Capacity</legend>
+                                                <div className="text-base font-medium text-gray-900" aria-hidden="true">
+                                                    Capacity
+                                                </div>
+                                                <p className="text-sm text-gray-500">Please select those that you could accomodate.</p>
+                                                <div className="mt-4 space-y-4">
+                                                    {/* Single */}
+                                                    <div className="relative flex items-start">
+                                                        <div className="flex h-5 items-center">
+                                                            <input
+                                                                id="capacity-single"
+                                                                name="capacity-single"
+                                                                type="checkbox"
+                                                                checked={data.canTakeSingles}
+                                                                onChange={e => setData('canTakeSingles', e.target.checked)}
+                                                                className="h-4 w-4 rounded border-gray-300 text-green-600 focus:ring-green-500"
+                                                            />
+                                                        </div>
+                                                        <div className="ml-3 text-sm">
+                                                            <label htmlFor="capacity-single" className="font-medium text-gray-700">
+                                                                Single
+                                                            </label>
+                                                        </div>
+                                                    </div>
+                                                    {/* Couple */}
+                                                    <div className="relative flex items-start">
+                                                        <div className="flex h-5 items-center">
+                                                            <input
+                                                                id="capacity-couple"
+                                                                name="capacity-couple"
+                                                                type="checkbox"
+                                                                checked={data.canTakeCouples}
+                                                                onChange={e => setData('canTakeCouples', e.target.checked)}
+                                                                className="h-4 w-4 rounded border-gray-300 text-green-600 focus:ring-green-500"
+                                                            />
+                                                        </div>
+                                                        <div className="ml-3 text-sm">
+                                                            <label htmlFor="capacity-couple" className="font-medium text-gray-700">
+                                                                Couple
+                                                            </label>
+                                                        </div>
+                                                    </div>
+                                                    {/* Family */}
+                                                    <div className="relative flex items-start">
+                                                        <div className="flex h-5 items-center">
+                                                            <input
+                                                                id="capacity-family"
+                                                                name="capacity-family"
+                                                                type="checkbox"
+                                                                checked={data.canTakeFamilies}
+                                                                onChange={e => setData('canTakeFamilies', e.target.checked)}
+                                                                className="h-4 w-4 rounded border-gray-300 text-green-600 focus:ring-green-500"
+                                                            />
+                                                        </div>
+                                                        <div className="ml-3 text-sm">
+                                                            <label htmlFor="capacity-family" className="font-medium text-gray-700">
+                                                                Family
+                                                            </label>
+                                                        </div>
+                                                    </div>
+                                                </div>
+                                            </fieldset>
+                                        </div>
+
+                                        <div className="pt-6">
+                                            {/* Can Take Pets? */}
+                                            <div className="sm:grid sm:grid-cols-2 sm:items-start sm:gap-4">
+                                                <label htmlFor="can-take-pets" className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2">
+                                                    Can you take pets?
+                                                </label>
+                                                <div className="mt-1 col-span-1 sm:mt-0">
+                                                    <select
+                                                        id="can-take-pets"
+                                                        name="can-take-pets"
+                                                        checked={data.canTakePets}
+                                                        onChange={e => setData('canTakePets', e.target.checked)}
+                                                        className="block w-full max-w-lg rounded-md border-gray-300 shadow-sm focus:border-green-500 focus:ring-green-500 sm:max-w-xs sm:text-sm"
+                                                    >
+                                                        <option value="true">Yes</option>
+                                                        <option value="false">No</option>
+                                                    </select>
+                                                </div>
+                                            </div>
+                                        </div>
+                                    </div>
+
+                                    {/* Actions */}
+                                    <div className="pt-5">
+                                        <div className="flex justify-end">
+                                            <button
+                                                type="button"
+                                                onClick={() => toggleFn(false)}
+                                                className="rounded-md border border-gray-300 bg-white py-2 px-4 text-sm font-medium text-gray-700 shadow-sm hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
+                                            >
+                                                Cancel
+                                            </button>
+                                            <button
+                                                type="submit"
+                                                disabled={processing}
+                                                className="ml-3 inline-flex justify-center rounded-md border border-transparent bg-green-600 py-2 px-4 text-sm font-medium text-white shadow-sm hover:bg-green-700 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
+                                            >
+                                                Create Offer
+                                            </button>
+                                        </div>
+                                    </div>
+                                </form>
+                            </Dialog.Panel>
+                        </Transition.Child>
+                    </div>
+                </div>
+            </Dialog>
+        </Transition.Root>
+    );
+}
+
+const CreateReportModal = ({ open, toggleFn }) => {
+    const mapRef = useRef();
+    const { auth } = usePage().props
+    const { data, setData, post, processing, errors } = useForm({
+        lat: -37.8136,
+        lng: 144.9631,
+        type: 'ROAD_DAMAGE',
+        user_id: auth.user.id
+    })
+
+    const onMapLoad = useCallback(() => {
+        mapRef.current.on('move', () => {
+            const coordinates = mapRef.current.getCenter();
+
+            setData(data => ({ ...data, lat: coordinates.lat }));
+            setData(data => ({ ...data, lng: coordinates.lng }));
+        });
+    }, []);
+
+    function submit(e) {
+        e.preventDefault()
+        post('/api/points-of-interest', {
+            onSuccess: () => toggleFn(false)
+        })
+    }
+
+    return (
+        <Transition.Root show={open} as={Fragment}>
+            <Dialog as="div" className="relative z-10" onClose={() => toggleFn(false)}>
+                <Transition.Child
+                    as={Fragment}
+                    enter="ease-out duration-300"
+                    enterFrom="opacity-0"
+                    enterTo="opacity-100"
+                    leave="ease-in duration-200"
+                    leaveFrom="opacity-100"
+                    leaveTo="opacity-0"
+                >
+                    <div className="fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity" />
+                </Transition.Child>
+
+                <div className="fixed inset-0 z-10 overflow-y-auto">
+                    <div className="flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0">
+                        <Transition.Child
+                            as={Fragment}
+                            enter="ease-out duration-300"
+                            enterFrom="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
+                            enterTo="opacity-100 translate-y-0 sm:scale-100"
+                            leave="ease-in duration-200"
+                            leaveFrom="opacity-100 translate-y-0 sm:scale-100"
+                            leaveTo="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
+                        >
+                            <Dialog.Panel className="relative transform overflow-hidden rounded-lg bg-white px-4 pt-5 pb-4 text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-sm sm:p-6">
+                                <form onSubmit={submit} className="space-y-6 divide-y divide-gray-200">
+                                    <div className="space-y-6 divide-y divide-gray-200">
+                                        <div>
+                                            <h3 className="text-lg font-medium leading-6 text-gray-900">Report</h3>
+                                            <p className="mt-1 text-sm text-gray-500">
+                                                We will display this on the map - thank you for helping keep us safe.
+                                            </p>
+                                        </div>
+
+                                        <div className="pt-6">
+                                            <legend className="sr-only">Location</legend>
+                                            <div className="text-base font-medium text-gray-900" aria-hidden="true">
+                                                Location
+                                            </div>
+                                            <div id='offer-location-select' className="h-64 w-full relative mt-4">
+                                                <Map
+                                                    ref={mapRef}
+                                                    initialViewState={{
+                                                        longitude: 144.9631,
+                                                        latitude: -37.8136,
+                                                        zoom: 5
+                                                    }}
+                                                    style={{
+                                                        width: '100%',
+                                                        height: '100%'
+                                                    }}
+                                                    onLoad={onMapLoad}
+                                                    mapStyle="mapbox://styles/mapbox/streets-v9"
+                                                    mapboxAccessToken="pk.eyJ1IjoiZGFuaWVsZmVyZ3Vzb24iLCJhIjoiY2w5YXFjazNtMGp1ZTNwcXdtMjBlYTc2YyJ9.2Cz8UmqgWB4VpagnJ6_ATw"
+                                                >
+                                                    <Marker longitude={data.lng} latitude={data.lat} anchor="bottom" />
+                                                    <GeolocateControl />
+                                                </Map>
+                                            </div>
+                                        </div>
+
+                                        <div className="pt-6">
+                                            {/* Hazard Type */}
+                                            <fieldset>
+                                                <legend className="sr-only">Hazard Type</legend>
+                                                <div className="text-base font-medium text-gray-900" aria-hidden="true">
+                                                    Hazard Type
+                                                </div>
+                                                <p className="text-sm text-gray-500">Please select the type of hazard.</p>
+                                                <div className="mt-4 space-y-4">
+                                                    {/* Single */}
+                                                    <select
+                                                        id="type"
+                                                        name="type"
+                                                        className="mt-1 block w-full rounded-md border-gray-300 py-2 pl-3 pr-10 text-base focus:border-green-500 focus:outline-none focus:ring-green-500 sm:text-sm"
+                                                        defaultValue="ROAD_DAMAGE"
+                                                        value={data.type}
+                                                        onChange={e => setData('type', e.target.value)}
+                                                    >
+                                                        <option value="ROAD_DAMAGE">Road Damage</option>
+                                                    </select>
+                                                </div>
+                                            </fieldset>
+                                        </div>
+                                    </div>
+
+                                    {/* Actions */}
+                                    <div className="pt-5">
+                                        <div className="flex justify-end">
+                                            <button
+                                                type="button"
+                                                onClick={() => toggleFn(false)}
+                                                className="rounded-md border border-gray-300 bg-white py-2 px-4 text-sm font-medium text-gray-700 shadow-sm hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
+                                            >
+                                                Cancel
+                                            </button>
+                                            <button
+                                                type="submit"
+                                                disabled={processing}
+                                                className="ml-3 inline-flex justify-center rounded-md border border-transparent bg-green-600 py-2 px-4 text-sm font-medium text-white shadow-sm hover:bg-green-700 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
+                                            >
+                                                Report Hazard
+                                            </button>
+                                        </div>
+                                    </div>
+                                </form>
                             </Dialog.Panel>
                         </Transition.Child>
                     </div>
@@ -124,13 +426,6 @@ const AuthModal = ({ open, toggleFn }) => {
                                         Github (Dev Only)
                                     </a>
                                     <a
-                                        href="/auth/redirect/apple"
-                                        className="inline-flex items-center justify-center rounded-md border border-gray-300 bg-white px-4 py-2 text-base font-medium text-gray-700 shadow-sm hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
-                                    >
-                                        <img className="-ml-1 mr-3 h-5 w-5 text-white" src='/assets/icons/apple.svg' />
-                                        Apple
-                                    </a>
-                                    <a
                                         href="/auth/redirect/facebook"
                                         className="inline-flex items-center justify-center rounded-md border border-gray-300 bg-white px-4 py-2 text-base font-medium text-gray-700 shadow-sm hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
                                     >
@@ -167,11 +462,146 @@ const AuthModal = ({ open, toggleFn }) => {
     )
 }
 
+const AccomodateTypeBadges = ({ offer }) => {
+    let badges = [];
+
+    if (offer.canTakeSingles && offer.canTakeCouples && offer.canTakeFamilies) {
+        return (
+            <span className="inline-flex items-center rounded bg-indigo-100 px-2 py-0.5 text-xs font-medium text-indigo-800">
+                <svg xmlns="http://www.w3.org/2000/svg" className="mr-1.5 h-3 w-3 text-indigo-400" viewBox="0 0 512 512"><path d="M470.6 105.4c12.5 12.5 12.5 32.8 0 45.3l-256 256c-12.5 12.5-32.8 12.5-45.3 0l-128-128c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L192 338.7 425.4 105.4c12.5-12.5 32.8-12.5 45.3 0z" /></svg>
+                Everyone
+            </span>
+        )
+    }
+
+    if (offer.canTakeSingles) {
+        badges.push(
+            <span className="inline-flex items-center rounded bg-indigo-100 px-2 py-0.5 text-xs font-medium text-indigo-800">
+                <svg xmlns="http://www.w3.org/2000/svg" className="mr-1.5 h-3 w-3 text-indigo-400" viewBox="0 0 512 512"><path d="M470.6 105.4c12.5 12.5 12.5 32.8 0 45.3l-256 256c-12.5 12.5-32.8 12.5-45.3 0l-128-128c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L192 338.7 425.4 105.4c12.5-12.5 32.8-12.5 45.3 0z" /></svg>
+                Singles
+            </span>
+        )
+    }
+
+    if (offer.canTakeCouples) {
+        badges.push(
+            <span className="inline-flex items-center rounded bg-indigo-100 px-2 py-0.5 text-xs font-medium text-indigo-800">
+                <svg xmlns="http://www.w3.org/2000/svg" className="mr-1.5 h-3 w-3 text-indigo-400" viewBox="0 0 512 512"><path d="M470.6 105.4c12.5 12.5 12.5 32.8 0 45.3l-256 256c-12.5 12.5-32.8 12.5-45.3 0l-128-128c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L192 338.7 425.4 105.4c12.5-12.5 32.8-12.5 45.3 0z" /></svg>
+                Couples
+            </span>
+        )
+    }
+
+    if (offer.canTakeFamilies) {
+        badges.push(
+            <span className="inline-flex items-center rounded bg-indigo-100 px-2 py-0.5 text-xs font-medium text-indigo-800">
+                <svg xmlns="http://www.w3.org/2000/svg" className="mr-1.5 h-3 w-3 text-indigo-400" viewBox="0 0 512 512"><path d="M470.6 105.4c12.5 12.5 12.5 32.8 0 45.3l-256 256c-12.5 12.5-32.8 12.5-45.3 0l-128-128c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L192 338.7 425.4 105.4c12.5-12.5 32.8-12.5 45.3 0z" /></svg>
+                Families
+            </span>
+        )
+    }
+
+    return badges;
+}
+
+const OffersPanelSection = ({ offers, setCreateModalOpen }) => {
+    if (offers.length === 0) {
+        return (
+            <div className="px-4 py-6">
+                <button
+                    type="button"
+                    onClick={() => setCreateModalOpen(true)}
+                    className="relative block w-full rounded-lg border-2 border-dashed border-gray-300 px-12 py-6 text-center hover:border-gray-400 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
+                >
+                    <PlusCircleIcon className="mx-auto h-12 w-12 text-gray-400" />
+                    <span className="mt-2 block text-sm font-medium text-gray-900">Create an offer</span>
+                </button>
+            </div>
+        )
+    }
+
+    return (
+        <div className="px-4 pt-6 pb-3">
+            <h2 className='text-lg font-medium'>Your offers</h2>
+            <ul role="list" className="divide-y divide-gray-200 text-sm">
+                {offers.map((offer) => (
+                    <li key={offer.id} className="py-4 grid gap-2">
+                        <div className='flex justify-between'>
+                            <div className="flex items-center">
+                                <HomeIcon className='h-5 w-5 mr-2' />
+                                <span>Offer of Accomodation</span>
+                            </div>
+                            <div className="flex items-center gap-2">
+                                <PencilSquareIcon className="h-3.5 w-3.5 text-gray-400 hover:text-gray-900 transition duration-75 cursor-pointer" />
+                                <MapPinIcon className="h-3.5 w-3.5 text-gray-400 hover:text-gray-900 transition duration-75 cursor-pointer" />
+                            </div>
+                        </div>
+                        <div className="flex items-center gap-2">
+                            <AccomodateTypeBadges offer={offer} />
+                            {offer.canTakePets && (
+                                <span className="inline-flex items-center rounded bg-indigo-100 px-2 py-0.5 text-xs font-medium text-indigo-800">
+                                    <svg xmlns="http://www.w3.org/2000/svg" className="mr-1.5 h-3 w-3 text-indigo-400" viewBox="0 0 512 512"><path d="M470.6 105.4c12.5 12.5 12.5 32.8 0 45.3l-256 256c-12.5 12.5-32.8 12.5-45.3 0l-128-128c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0L192 338.7 425.4 105.4c12.5-12.5 32.8-12.5 45.3 0z" /></svg>
+                                    Pets
+                                </span>
+                            )}
+                        </div>
+                    </li>
+                ))}
+            </ul>
+        </div>
+    )
+}
+
+const ReportsPanelSection = ({ reports }) => {
+    if (reports.length === 0) {
+        return;
+    }
+
+    return (
+        <div className="px-4 pt-6 pb-3">
+            <h2 className='text-lg font-medium'>Your reports</h2>
+            <ul role="list" className="divide-y divide-gray-200 text-sm">
+                {reports.map((report) => (
+                    <li key={report.id} className="py-4 grid gap-2">
+                        <div className='flex justify-between'>
+                            <div className="flex items-center">
+                                <ExclamationTriangleIcon className='h-5 w-5 mr-2' />
+                                <span>Report</span>
+                            </div>
+                            <div className="flex items-center gap-2">
+                                <XMarkIcon className="h-3.5 w-3.5 text-gray-400 hover:text-gray-900 transition duration-75 cursor-pointer" />
+                                <MapPinIcon className="h-3.5 w-3.5 text-gray-400 hover:text-gray-900 transition duration-75 cursor-pointer" />
+                            </div>
+                        </div>
+                    </li>
+                ))}
+            </ul>
+        </div>
+    );
+}
+
+const LoginPanelSection = ({ toggleFn }) => {
+    return (
+        <div className="px-4 grid gap-3 py-6">
+            <h2 className='text-lg font-medium'>Register an offer</h2>
+            <p className='text-sm font-normal'>Please login to create an offer.</p>
+            <button
+                type="button"
+                onClick={() => toggleFn(true)}
+                className="inline-flex items-center rounded-md border border-transparent bg-green-100 px-3 py-2 text-sm font-medium leading-4 text-green-700 hover:bg-green-200 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
+            >
+                Login
+            </button>
+        </div>
+    );
+}
+
 export default function SidebarLayout({ children }) {
     const [sidebarOpen, setSidebarOpen] = useState(false)
     const [authModalOpen, setAuthModalOpen] = useState(false);
-    const [createModalOpen, setCreateModalOpen] = useState(true)
-    const { auth } = usePage().props
+    const [createOfferModalOpen, setCreateOfferModalOpen] = useState(false);
+    const [createReportModalOpen, setCreateReportModalOpen] = useState(false);
+    const { auth, offers, reports } = usePage().props
 
     const logoutFn = () => {
         Inertia.visit('/logout', {
@@ -181,8 +611,9 @@ export default function SidebarLayout({ children }) {
 
     return (
         <div className='text-gray-900'>
-            {/* Create modal */}
-            <CreateModal open={createModalOpen} toggleFn={setCreateModalOpen} />
+            {/* Modals */}
+            <CreateOfferModal open={createOfferModalOpen} toggleFn={setCreateOfferModalOpen} />
+            <CreateReportModal open={createReportModalOpen} toggleFn={setCreateReportModalOpen} />
 
             {/* Auth modal */}
             <AuthModal open={authModalOpen} toggleFn={setAuthModalOpen} />
@@ -241,20 +672,41 @@ export default function SidebarLayout({ children }) {
                                             alt="Helping Homes"
                                         />
                                         <h1 className='font-brand text-green-600'>Helping Homes</h1>
-
                                     </div>
-                                    <nav className="mt-5 space-y-1 px-2">
-                                        {/* TODO: Navigation */}
+                                    <div className='px-4 mt-2'>
+                                        <p className='text-gray-400 italic text-xs'>Proudly by Helping Group</p>
+                                    </div>
+                                    <nav className="flex-1 space-y-1 divide-y">
+                                        {/* TODO */}
+                                        {/* Link to give page */}
+                                        <div className="px-4 grid gap-3 py-6">
+                                            <h2 className='text-lg font-medium'>Other ways to help</h2>
+                                            <p className='text-sm'>Not everyone can offer accomodation, however there are other ways to make an immediate, meaningful impact in this time of crisis.</p>
+                                            <p className='text-sm'>Each of these organisations are helping communities affected by natural disaster right now - <b>and they need our help.</b></p>
+                                            <a
+                                                href="/help"
+                                                className="inline-flex items-center rounded-md mt-2 border border-gray-300 bg-white px-3 py-2 text-sm font-medium leading-4 text-gray-700 shadow-sm hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
+                                            >
+                                                Find out how
+                                            </a>
+                                        </div>
+                                        {/* Assistance */}
+                                        <div className="px-4 grid gap-3 py-6">
+                                            <h2 className='text-lg font-medium'>Get assistance</h2>
+                                            <p className='text-sm text-red-600 font-bold'>If you need urgent assistance, please call 000 immediately.</p>
+                                            <p className='text-sm'>There are other fantastic organisations who are helping in so many ways, and we want to recognise them, and enable people find the help they need.</p>
+                                        </div>
                                     </nav>
                                 </div>
                                 {auth.user && (
                                     <div className="flex flex-shrink-0 border-t p-4">
-                                        <button onClick={() => logoutFn()} className="group block flex-shrink-0 ml-2">
-                                            <p className="text-base font-medium ">{auth.user.name}</p>
-                                            <p className="text-sm font-medium text-gray-400 group-hover:">Logout</p>
+                                        <button onClick={() => logoutFn()} className="group block w-full flex-shrink-0 ml-2">
+                                            <p className="text-sm font-medium ">{auth.user.name}</p>
+                                            <p className="text-xs font-medium text-gray-400 group-hover:">Logout</p>
                                         </button>
                                     </div>
                                 )}
+
                             </Dialog.Panel>
                         </Transition.Child>
                         <div className="w-14 flex-shrink-0" aria-hidden="true">
@@ -265,7 +717,7 @@ export default function SidebarLayout({ children }) {
             </Transition.Root>
 
             {/* Static sidebar for desktop */}
-            <div className="hidden md:fixed md:inset-y-0 md:flex md:w-64 md:flex-col">
+            <div className="hidden md:fixed md:inset-y-0 md:flex md:w-72 md:flex-col">
                 {/* Sidebar component, swap this element with another sidebar if you like */}
                 <div className="flex min-h-0 flex-1 flex-col">
                     <div className="flex flex-1 flex-col overflow-y-auto pt-5 pb-4">
@@ -281,34 +733,15 @@ export default function SidebarLayout({ children }) {
                             <p className='text-gray-400 italic text-xs'>Proudly by Helping Group</p>
                         </div>
                         <nav className="flex-1 space-y-1 divide-y">
-                            {/* Register to sign up */}
-                            {!auth.user && (
-                                <div className="px-4 grid gap-3 py-6">
-                                    <h2 className='text-lg font-medium'>Register an offer</h2>
-                                    <p className='text-sm font-normal'>Please login to create an offer.</p>
-                                    <button
-                                        type="button"
-                                        onClick={() => setAuthModalOpen(true)}
-                                        className="inline-flex items-center rounded-md border border-transparent bg-green-100 px-3 py-2 text-sm font-medium leading-4 text-green-700 hover:bg-green-200 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
-                                    >
-                                        Login
-                                    </button>
-                                </div>
-                            )}
-                            {/* Submit an offer */}
-                            {auth.user && (
-                                <div className="px-4 py-6">
-                                    <button
-                                        type="button"
-                                        className="relative block w-full rounded-lg border-2 border-dashed border-gray-300 p-12 text-center hover:border-gray-400 focus:outline-none focus:ring-2 focus:ring-green-500 focus:ring-offset-2"
-                                    >
-                                        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth="1.5" stroke="currentColor" className="mx-auto h-12 w-12 text-gray-400">
-                                            <path strokeLinecap="round" strokeLinejoin="round" d="M12 9v6m3-3H9m12 0a9 9 0 11-18 0 9 9 0 0118 0z" />
-                                        </svg>
-                                        <span className="mt-2 block text-sm font-medium text-gray-900">Create an offer or report</span>
-                                    </button>
-                                </div>
-                            )}
+                            {/* Login */}
+                            {auth.user === false && <LoginPanelSection toggleFn={setAuthModalOpen} />}
+
+                            {/* Offers */}
+                            {auth.user && <OffersPanelSection offers={offers} setCreateModalOpen={setCreateOfferModalOpen} />}
+
+                            {/* Reports */}
+                            {auth.user && <ReportsPanelSection reports={reports} setCreateModalOpen={setCreateReportModalOpen} />}
+
                             {/* Link to give page */}
                             <div className="px-4 grid gap-3 py-6">
                                 <h2 className='text-lg font-medium'>Other ways to help</h2>
@@ -340,7 +773,7 @@ export default function SidebarLayout({ children }) {
                     )}
                 </div>
             </div >
-            <div className="flex flex-1 flex-col md:pl-64 h-screen">
+            <div className="flex flex-1 flex-col md:pl-72 h-screen">
                 <div className="sticky top-0 z-10 p-1 md:hidden">
                     <button
                         type="button"